+++ Package.swift
@@ -16,7 +16,7 @@ let package = Package(
     dependencies: [
       .package(
         url:"https://github.com/apple/swift-argument-parser",
-        .exact("0.0.6")),
+        .exact("0.2.0")),
       .package(url: "https://github.com/erica/Swift-General-Utility", from: "0.0.1"),
     ],
     targets: [