+++ xcopen/main.swift
@@ -5,101 +5,103 @@ import ArgumentParser
 import GeneralUtility
 
 struct Xcopen: ParsableCommand {
-    static var configuration = CommandConfiguration(helpNames: [.short, .customLong("help", withSingleDash: true)])
-    
-    @Flag(help: .hidden) // provides undocumented --help
-    var help = false
-    
-    @Argument(help: "Files to open. Leave blank to search for xcodeproj and xcworkspace in the working directory")
-    var files: [String] = []
-    
-    @Flag(name: [.short, .customLong("all", withSingleDash: true)], help: "In addition, open Package.swift")
-    var allXcodeTypes = false
-    
-    @Flag(name: [.short, .customLong("docs", withSingleDash: true)], help: "Open all .md and .txt files")
-    var docTypes = false
-    
-    @Flag(name: [.short, .customLong("new", withSingleDash: true)], help: "Create new files and then open them in Xcode")
-    var newFiles = false
-    
-    @Flag(name: [.short, .customLong("focus", withSingleDash: true)], help: .hidden) // "Set Xcode focus on the first file"
-    var focusXcode = false
-    
-    @Flag(name: [.short, .customLong("pkg", withSingleDash: true), .customLong("package", withSingleDash: true), .customLong("edit", withSingleDash: true)], help: .hidden)
-    var openPackageInTextEdit = false
-    
-    @Flag(name: [.customShort("b"), .customLong("bg", withSingleDash: true), .customLong("background", withSingleDash: true)], help: "Open Xcode in the background")
+    static var configuration = CommandConfiguration(discussion: """
+    xcopen <files>...        Open files in Xcode.
+    xcopen                   Open workspace or xcodeproj.
+    xcopen ws                Only search for workspace.
+    xcopen xc                Only search for xcodeproj.
+    xcopen docs              Open .md and .txt files.
+    xcopen pkg               Open Package.swift in TextEdit.
+    xcopen xpkg              Open Package.swift in Xcode.
+    xcopen new               Create new files (if they don't exist) and
+                             open in Xcode.
+    """)
+
+
+    @Argument(help: "Files to open. If blank, opens xcworkspace or, if not found, searches for xcodeproj.")
+    var paths: [String] = []
+
+    @Flag(name: [.customShort("b"), .customLong("bg", withSingleDash: true), .customLong("background")], help: "Open Xcode in the background")
     var openInBackground = false
-    
+
     func run() throws {
-        guard !help
-            else { throw CleanExit.helpRequest() }
-        
-        // Little undocumented feature for my own use. Sort of subcommands
-        // without actually being subcommands.
         var isDir: ObjCBool = false
-        if files.count == 1 {
-            switch files[0] {
+
+        // Handle the empty paths case first. Everything else references the paths list.
+        guard !paths.isEmpty else {
+            let hasWorkspace = try !FileManager.default.contentsOfDirectory(atPath: ".")
+                .filter({ $0.hasSuffix(".xcworkspace") }).isEmpty
+            switch hasWorkspace {
+            case true:
+                try Utility.searchAndOpen([".xcworkspace"], bg: openInBackground)
+            case false:
+                try Utility.searchAndOpen([".xcodeproj"], bg: openInBackground)
+            }
+            return
+        }
+
+        // Subcommands without being subcommands, allowing the utility
+        // to be called without any arguments
+        if paths.count == 1 {
+            switch paths[0] {
+
+            // Open a workspace if there's one here
             case "ws":
-                // Just open a workspace if there's one here
-                try searchAndOpen([".xcworkspace"], bg: openInBackground)
+                try Utility.searchAndOpen([".xcworkspace"], bg: openInBackground)
                 return
+
+            // Opens all xcodeproj
+            case "xc":
+                try Utility.searchAndOpen([".xcodeproj"], bg: openInBackground)
+                return
+
+            // Open docs
             case "docs":
-                // Just open docs
-                try searchAndXcodeOpen([".txt", ".md"], bg: openInBackground)
+                try Utility.searchAndXcodeOpen([".txt", ".md"], bg: openInBackground)
                 return
+
+            // Open Package.swift in Textedit
             case "pkg":
-                // Just open Package.swift in Textedit
-                try openPkgInTextEdit()
+                try Utility.openPkgInTextEdit()
                 return
+
+            // Open Package.swift in Xcode
             case "xpkg":
-                // Just open Package.swift in Xcode
-                try searchAndXcodeOpen(["Package.swift"], bg: openInBackground)
+                try Utility.searchAndXcodeOpen(["Package.swift"], bg: openInBackground)
                 return
-            case _ where FileManager.default.fileExists(atPath: files[0], isDirectory: &isDir) && isDir.boolValue == true :
-                let dirURL = URL(fileURLWithPath: files[0])
-                var filesToOpen: [String] = []
-                for file in try FileManager.default.contentsOfDirectory(atPath: files[0]) {
-                    let path = dirURL.appendingPathComponent(file).path
-                    _ = FileManager.default.fileExists(atPath: path, isDirectory: &isDir)
-                    if isDir.boolValue == true { continue }
-                    filesToOpen.append(path)
-                }
-                _ = try Utility.execute(commandPath: "/usr/bin/open", arguments: (openInBackground ? ["-g"] : []) + ["-a", xcurl().path] + filesToOpen)
+
+            // This is handled below
+            case "new":
+                break
+
+            // Check whether there is a single path and it is a directory
+            case _
+                    where
+                    paths.count == 1
+                    && FileManager.default.fileExists(atPath: paths[0], isDirectory: &isDir)
+                    && isDir.boolValue == true :
+                let dirURL = URL(fileURLWithPath: paths[0])
+                let filesToOpen = try FileManager.default
+                    .contentsOfDirectory(atPath: paths[0])
+                    .compactMap({ path -> String? in
+                        let filePath = dirURL.appendingPathComponent(path).path
+                        _ = FileManager.default.fileExists(atPath: path, isDirectory: &isDir)
+                        return isDir.boolValue == true ? nil : filePath
+                    })
+                _ = try Utility.execute(commandPath: "/usr/bin/open", arguments: (openInBackground ? ["-g"] : []) + ["-a", Utility.xcurl().path] + filesToOpen)
                 return
             default:
                 break
             }
         }
-        
-        if openPackageInTextEdit {
-            try openPkgInTextEdit()
+
+        // Special case "new" as the first file mentioned
+        if paths[0] == "new" {
+            try Utility.createNewFilesAndOpen(Array<String>(paths.dropFirst()), bg: openInBackground)
             return
         }
-        
-        if newFiles {
-            try createNewFilesAndOpen(files, bg: openInBackground)
-            return
-        }
-        
-        if docTypes {
-            try searchAndXcodeOpen([".txt", ".md"], bg: openInBackground)
-            return
-        }
-        
-        if focusXcode {
-            guard !files.isEmpty
-                else { throw RuntimeError("Focus requires a file name") }
-            try searchAndOpen([".xcodeproj", ".xcworkspace"] + (allXcodeTypes ? ["Project.swift"] : []), bg: openInBackground)
-            try focus(files[0])
-            return
-        }
-        
-        if files.isEmpty {
-            try searchAndOpen([".xcodeproj", ".xcworkspace"] + (allXcodeTypes ? ["Project.swift"] : []), bg: openInBackground)
-        } else {
-            try searchAndXcodeOpen(files, bg: openInBackground)
-        }
+
+        try Utility.searchAndXcodeOpen(paths, bg: openInBackground)
     }
 }